refactor: colocate deck_results_filter as deck_research_panel.results_filter
The pure (wx-free) deck-results filter logic had only deck-research UI
consumers. Move it next to them as
widgets/panels/deck_research_panel/results_filter.py per the architecture's
colocation principle. Make the package __init__ lazy-load the wx panel
(mirroring widgets/panels/__init__.py) so results_filter and its headless
unit test stay importable without wx.
