+++ widgets/panels/deck_research_panel/frame/filters.py
@@ -11,7 +11,7 @@ from typing import TYPE_CHECKING
 import wx
 
 from utils.constants import PADDING_MD
-from utils.deck_results_filter import PLACEMENT_FIELDS, PLACEMENT_OPERATORS
+from widgets.panels.deck_research_panel.results_filter import PLACEMENT_FIELDS, PLACEMENT_OPERATORS
 from widgets.panels.deck_research_panel.frame.centered_choice import _CenteredChoice
 from widgets.stylize import stylize_button, stylize_choice, stylize_label, stylize_textctrl
