+++ widgets/panels/deck_research_panel/results_filter.py
